fixed writing shared resharper settings
obviously the new ReSharper tries to load the relative path first and then the absolute path. the absolute path is now obsolete but it doesn't hurt as we set it with every gd. the crucial part is the relative path which was pointing to the old location hence being empty.