Update existing developer tools
Interim solution for updating the Aderant Developer Tools when they are already installed. If we want to "publish" a new version, we just drop a build of Aderant.DeveloperTools.vsix into the Build.Tools directory and commit that to git. Next time a console opens we pull the latest Build.Infrastructure and a version check will take place.