PR 5117: Revert build caching
 - Revert build caching

This reverts using a shared Build Infrastructure/devtools folder on the CI box. The problem is when a devtools update occurs all new builds try to update the cache which is impossible as other builds are using it and so the new build fails.
Rather than fix this, we'll just back it out.

Related work items: #169925