Merged PR 16892: Added a switch to the Build-ExpertPatch command
Added a switch to the Build-ExpertPatch command to differentiate between building pre 803 and 803 or later patches.
The patchbuilder folders are now split between 'current' and 'pre803' for pre 803 lines. This removes the requirement for us to rebuild the prerefactor branch for old lines, and gives us a split where the patchbuilder (and db project) are compatible.

Related work items: #179676