Merged PR 22870: Pull buildscripts.zip from master build.Infrastructure
This is a pretty janky workaround to pull the buildscripts.zip from the master branch of build.infrastructure instead of the current version of sequencer that everyone is using. I take advantage of the rest api to download the folder as a zip then manipulate it to suit our needs.
Once were happy with this, there will be an additional step to update the build.proj of the framework repo.

Any feedback is welcome, I'm surprised this works tbh

Related work items: #208862