Merged PR 23522: removing check to see if project is web project when calculating files to
removing check to see if project is web project when calculating files to restore. This fixes missing web zips which are depended on in their corresponding bin/module folders. The issue is a result of not adding dependencies into the tracked projects list in the pipeline service, we have decided not to add them since this is an expensive operation.