Merged PR 24084: Fix an issue where we allowed a pre node to be in the same group as a project
Fix an issue where we allowed a pre node to be in the same group as a project that depends on the outputs of the pre node. This leads to a race condition where we may not restore objects fast enough and so a build can fail.

This fixes an issue where Collections Outlook addin depends on the Aderant.Outlook.Extensions.
While the dependency order is correct with the previous implementation when we remove Aderant.Outlook.Extensions from the graph (as it can come from the cache) we don't swap in a dependency on the download operation of the provider of Aderant.Outlook.Extensions which leads to a race condition.