Merged PR 24131: This adds tracking support to MSI projects so don't get cache misses.
This adds tracking support to MSI projects so don't get cache misses.

This fixes these harmless warnings which just means we always need to rebuild the MSIs as we can never find them in the build cache as we never added .wixproj to the list of tracked projects.

![image.png](http://tfs:8080/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/24131/attachments/image.png)

Related work items: #172389