Merged PR 24344: Merge ignore-merge-commits to 187604-Sequencer
Teach the version control service to ignore merge commits. We only want the common ancestor to be another branch if possible.

Prior to the change.

(command: (Get-SourceTreeMetadata).Changes)

![image.png](http://tfs:8080/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/24344/attachments/image.png)

After the change.

![image.png](http://tfs:8080/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/24344/attachments/image%20(2).png)

The problem is the logic followed the wrong parent when it encountered a merge commit, so the common ancestor was set to itself and things went south from there.

Related work items: #213968