Merged PR 24523: Merge environment_xml_finder to 187604-Sequencer
Search the branch local directory first (e.g. C:/git/master) if the environment.xml exists. If it exists, set this path as the binaries directory.

If environment.xml does not exist in the local directory, append Binaries to this path and set it as the binaries directory.

Related work items: #209095