Merged PR 25516: Assume the inputs are always out of date.
When the build cache is not used for a project which is tracking inputs then assume the inputs are always out of date.

Previously if we did not have a state file (due to no cache) then we could not tell if the dependency was up to date - we assumed it was.
This change flips that and if we are tracking inputs and no cache is present assumes that the dependency is always out of date.

This defeats the CachedBuildNotFound optimization where if a dependency is flagged as only 'CachedBuildNotFound' then we do mark downstream dependencies as affected.

Now the dependency will also be marked as "InputsChanged' so the tree analyser will build a graph that includes the stale dependency. This will result in over building if the input file track list is promiscuous but is more correct.