Merged PR 26880: Add metadata to determine when to use hard links for packages
Add metadata to determine when to use hard links for packages.

Unless this is set, the default value of true will be used for copying files for packages.

This change is required as we are packaging Microsoft assemblies from Visual Studio, and we need to copy them as opposed to creating hard links as they cannot be cleaned successfully as the assemblies can be locked by MSBuild.

Related work items: #224050