Merged PR 28859: BUGFIX: Delete file should delete a file if the file exists.
BUGFIX: Delete file should delete a file if the file exists.

This resolves an issue where you can get compiler failures if you have old stuff in the **_as** folder which doesn't get replaced when restoring objects as part of ```RunRetrieveArtifacts``` from the build cache as we never replace the target file.

Eg this target - the log statement was a lie and things where not replicated to **_as**

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/28859/attachments/image%20(2).png)

Example compiler failure as the version of the assembly in the reference path has out of date type info.
![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/28859/attachments/image.png)

Related work items: #229776