Merged PR 29810: Update the packaging process to produce web packages.
Update the packaging process to produce web packages.
If the assemblies contain "Web", package the assemblies into the web packages and mark the package as a test package.

Also testing the **Optional Build System** (so that we can invoke Expert-Test build using the custom PR branch in Build.Infrastructure).

Related work items: #232230, #232561