Merged PR 30277: Merge RoslynRuleIntegrationTest to master
Contains fixes for a number of Roslyn Rule related bugs, in addition to restructuring how the rules are loaded and tested.

This should have no impact on how the rules themselves run, but does allow us to be effectively test them.

Also includes a number of formatting updates to remove redundant files and classes.

Related work items: #208983, #212063, #215648, #231771