Merged PR 31299: Don't use custom-tsconfig-->tsc build target when Microsoft.TypeScript.MSBuild is present
**Issue:**
Going forward we want to use the MSBuild target for TypeScript from NuGet. When we do so, we cannot rely on tsc.exe to be available in the ProgramFiles\Microsoft SDKs\TypeScipt directory.

**Solution:**
Don't use custom-tsconfig-->tsc build target when Microsoft.TypeScript.MSBuild is present, i.e. when we have download this from nuget and copied into the packages folder.

Related work items: #225523