Merged PR 31590: Enable deterministic builds
Enable deterministic builds.

To allow for deterministic workflow builds we need to prevent the build engine from updating the debug symbol element as the element value does not respect path map. This change teaches the system to remove the debugger extension that rewrites the debug symbol element to ensure that the XAMLX in the obj directory is identical given the same inputs.

Related work items: #235872, #235897