Merged PR 31736: Ignore imports that are invalid if possible to allow for analysis before
Ignore imports that are invalid if possible to allow for analysis before dependencies are available.
If we do not have the WiX targets for example we cannot load the project as the props/targets file are not available yet but we still want to load the project.

Required to build Marketing.

Related work items: #123065