Merged PR 31852: FEATURE: Add the ability to declare an ItemGroup that provides custom test settings
FEATURE: Add the ability to declare an ItemGroup that provides custom test settings. You can do this by adding an ItemGroup of type 'RunSettings' with the include being the XML path the Value metadata being the value

eg
```
<ItemGroup>
 <RunSettings Include="/RunSettings/MSTest/DeleteDeploymentDirectoryAfterTestRunIsComplete">
        <Value>false</Value>
 </RunSettings>
</ItemGroup>
```
BUGFIX: Restore bm -Target RunTests (was broken as the target was renamed)
BUGFIX: Duplicate PowerShell logging when invoking Paket
BUGFIX: No console output when using nested PowerShell when using the PowerShellScript task
BUGFIX: Wire up CTRL+C to actually kill TestHost and VSTest
CLEAN UP: Remove legacy loggers and tasks assemblies that are no longer used.

Almost removed 'AssemblyPlatformDataKey' which was used to copy item group metadata around. This was supposed to be to determine assembly platforms within RunTest.ps1 but this was dropped so we do not need this data anymore and so has been removed.

#235206

Related work items: #235206