Merged PR 34481: Add default patterns for Build Cache Zip restoration
Currently a lot of modules have dir.props files which use one of two patterns (Web. * .zip or Aderant. * .Api.zip) to find build cache zips to restore. We would like to introduce these as default patterns - a dir.props file can say that it wants the default patterns by including the property RestorePackagedFilesFromBuildCacheZips. We are going to add a new dir.props file for Inquiries in master, and we intend to convert other dir.props to use this default option as appropriate. Anything that does not explicitly ask for the default will continue to function as it currently does.

Related work items: #244418