Merged PR 34631: Merge relative-path-to-paket-files to master
These changes are so I can do this and introduce a new top level link. As the dependency link system was encoded in two places previously it was unsatisfactory to duplicate the mapping information for a third directory in two places.
![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/34631/attachments/image.png)

Refactor of GetDependencies.

As GetDependencies was offered as an external public function the symlink map and dependency directory information can no longer be encoded into the build but instead needs to be made available from some other information source.
This change reads the existing 'Branch Config' which has a non-relative directory segment that details the desired shared dependency directory and applies that to the build. This now makes the build and 'gd' consistent.

Moves all duplicated knowledge between the MSBuild engine and the PS variant of GetDependencies into the class that was designed to be the orchestrator - ResolverWorkflow.

The symlink map is now created by the resolver, rather than the build as this is the only class that is common between the two entry points.

Introduces a new task.
Introduces BuildTestRunParametersXml that allows directories to proffer run settings to the build.

The duplication of the discovery of 'BranchConfig' has been simplified and turned into a common function between PowerShell and C#.

Related work items: #245415