Merged PR 34846: Merge upgrade-paket to master
Upgrades Paket to the latest stable version.

Adds a new command switch ```-SkipNugetPackageHashCheck``` to allow by passing of the NuGet package hash check. This is useful when messing around with dependencies but you don't want the system to disallow you from using a cached build as you know your changes are safe.

Performance optimization - fixes a minor issue when writing the single dependency file out where the TFM was not preserved which then allowed for downloading of Linux/OSX packages for items in groups other than main which is silly as we build for Windows only.

Related work items: #236254