Merged PR 35177: Merge build/package-hash-version-exclusions to master
Improve artifact selection process.

## How to exclude package versions from the hashing process
In the ```Build\BranchConfig.xml``` file, one can add the following section:
```XML
<ArtifactMetadataConfiguration>
  <PackageHashVersionExclusions>Aderant.Libraries.Models;Aderant.StoredProcedures</PackageHashVersionExclusions>
</ArtifactMetadataConfiguration>
```
This will exclude the versions for ```Aderant.Libraries.Models``` and ```Aderant.StoredProcedures``` when calculating the package hash.

This is useful for packages where we are tracking the contents via the ```TrackedFiles``` ItemGroup in the ```dir.props``` file.

### Changes
* Enable excluding package versions from lock file hashing alorithm
* Consider package hash during change file tracking process
* Attempt to find an exact artifact match based based on change file tracking

Related work items: #245468