Merged PR 37130: Add support for shared projects.
Add support for shared projects.

We do not considered shared projects a duplicate as they are intended to be shared.
Tested by applying the change manually to my build on SVBLD305 and running the server build interactively.

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/37130/attachments/image.png)

Does not affect the build plan as we do not consider shproj(s) for building.

#251694

Related work items: #251694