Merged PR 37703: Improve performance of versioning targets by not reading the lock file for
Improve performance of versioning targets by not reading the lock file for every project.

Instead we ask CSC to read the lock file and also attempt to guide the targets to a exiting file in the outer project. If this fails we fall back to finding the file ourselves.

Lock file is also moved to an embedded resource.

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/37703/attachments/image.png)

Related work items: #247296