Merged PR 37761: Update vs command to respect current directory
Previously the vs command was driven by ```cm``` which provided a confusing end-user experience where it was possible to run ```cm```, change directory and then use the ```vs``` command which would open a solution in the previous directory.

The behavior of the ```vs``` command has been updated to open the solution in the current directory, or subdirectory matching the provided module name.

Related work items: #253460