Merged PR 38597: Improve performance by reducing logging out large collections
Improve performance by reducing logging out large collections.

If you think about what this does, it's doing a ```string.Join()``` on ```@(MyItems)``` to compare if it matches an empty string. Luckily MSBuild has a "fast path" for evaluating against an empty string, but it still can generate the log message:

```
Target "Foo" skipped, due to false condition; ('@(MyItems)' != '') was evaluated as ('A;B;C' != '')
```

If ```@(MyItems)``` was 100 full paths to files, this would be a long log message!

This change uses a fast path on both the logging and the condition.