Merged PR 38675: Add artifact packaging for API assemblies
**Add artifact packaging for API assemblies**

**Problem:**

API integration tests are failing on the release pipeline because they cannot find the corresponding API assembly. Eg IntegrationTest.MatterManagement.Api cannot find Aderant.MatterManagement.Api.

The pipeline sets up symlinks to ExpertSource so that we can find dependencies, but the Apis are treated like Web modules, so instead of having the assembly in ExpertSource you have the website zip file, so you cannot find the dependency in ExpertSource.

**Solution:**

For Web modules we have a special step in the build where we package the Web assemblies alongside the tests, so that the tests can find them. This PR does the same thing for APIs so that their tests can find their dependencies.

**Notes:**

The Time API tests don't have this issue, because they output to bin\module like a normal assembly, instead of to bin like a website. If this goes in we can tidy that up so that it behaves the same as the other APIs.

It seems that on some days the MatterManagement API tests mostly work on the pipeline, because they find the assembly in an unexpected location, the parent directory of ExpertSource. But some tests still fail, seemingly because they load two copies of the same type from copies of the assembly in two different locations into the same app domain. Possibly there's some machine which needs some tidy-up.

Related work items: #255459