Merged PR 38760: Don't traverse directories that do not exist.
Don't traverse directories that do not exist.

This resolves an issue where if you change branches you may end up with directories that are only partially valid and so the directory contribution is valid but the path within that directory does not exist. If you then run ```gd -update``` you may get an exception 'Path Not Valid' as we try to look into a path that doesn't exist.

By design the scanner fails on all errors.  This change relaxes that design.