Merged PR 38823: Performance improvements for attribute generation
Performance improvements for attribute generation.

When generating the AssemblyInfo.cs file we repeat the same work for calculating the new assembly version all projects within a cone. We also then repeat the same work for calculating any additional AssemblyMetadata attributes.

This change updates a cloned copy of CommonAssemblyInfo.cs once, and then shares that between all projects using the same original CommonAssemblyInfo.cs.

Optimizations to the enrichment of AssemblyInfo.g.cs which is the per-project file.
Now we only write the g.cs file if there are attributes that need to be applied at the project level such as integration test attributes. Previously we tried to do this with no regard if there was actually work to do which wasted a few hundred milliseconds for each project.

  There are at least 3 AssemblyInfo.cs files involved.
  1) \<Temp>\CommonAssemblyInfoFile.cs    = A clone of CommonAssemblyInfoFile.cs with the copyright date modified (subject to race condition)
  2) \<Temp>\CommonAssemblyInfoFile.g.cs  = Repository metadata file (git version etc) (subject to race condition)
  3) \obj\AssemblyInfo.gs.cs              = A settings that apply to an individual project

Also improved the speed of the role file finder.

Related work items: #247296