Merged PR 39007: Force UpdateCommonAssemblyInfo to happen after SetBuildNumbers
Assemblies were not getting a build version number stamped even though we were correctly assigning one. This was caused by the target that should stamp the assembly being called before we have deduced the version number.

This change enforces the targets to run in the correct order

Related work items: #256128