Merged PR 39049: Do not expand unnecessarily
Do not expand unnecessarily.

A check was added to confirm that all modules in a DependencyManifest.xml also exist as directories on the computer. This conflicts with a previous design intent to _not_ check for this case. The check causes build failures.

The code documentation states that the module referenced by DependencyManifest will be added to a list and processed later - deferred processing. However the check  in ```ValidatePaths``` breaks this behaviour as it assumes that all the references are nested in the same repository as the file being analyzed.

This change stops the check from working on all repositories that only have a single level such as Deployment, Internal, SDK.Database etc. Going forward we could probably just delete  ```ValidatePaths```.

This change prevents an exception building the mentioned repositories

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/39049/attachments/image.png)

@<ADERANT_AP\jaishree.kulkarni> @<ADERANT_AP\robert.gray> @<ADERANT_AP\anthony.prisk>

Related work items: #249839