Merged PR 39061: BUGFIX: Fixed an issue where the legacy folder based dependency fetch was broken
BUGFIX: Fixed an issue where the legacy folder based dependency fetch was totally broken.
Now the fetch checks if the manifest has a DefaultSource and adds that to the list of sources to fetch from.
There was an issue where the folder resolver was not enabled unless the NuGet resolver was also enabled.

@<ADERANT_AP\jaishree.kulkarni>