Merged PR 39063: Partition the CommonAssemblyInfo.cs file by worker ID
Adds the current worker ID items to hash to produce a deterministic yet process dependent value to minimize but not eliminate generating a hash used by another worker.

This resolve an issue where two nodes can generate the same temp path which causes a build break due to 'file in use'.

- Improve shell loading speed. Found a bug in InProcessJobs that caused the main thread to block which stopped the loading of VS environment variables being asynchronous.
Rewrote InProcessJobs to use a runspace in PowerShell directly to avoid having to compile a separate DLL for threading which greatly simplifies the code and has the benefit of speeding up the first start.

- Upgrade paket to the latest version.
- Migrate Dataflow to match the version used by MSBuild 16.0
- Migrate Compiler.Unsafe to match the version used by MSBuild 16.0

Related work items: #256475