Merged PR 39187: Null check the fusion logging registry entry
We are checking the fusion logging registry entry to make sure that fusion logging is disabled. Not all of the build servers have this registry entry (seems some of them don't have fuslog installed), so we need to handle the possibility of the registry entry not existing.

Related work items: #256475