Merged PR 39250: Fix up scheduled tasks
Simplifies the scheduled task set for agents.

The driving force is there are 3 clean up tasks that run at start up which were concurrent with the agent setup script which leads to a classic race condition. Now there is a single startup task that does a cleanup at start up and then installs the agent.

Fixes a bug with the previous task set - it was trying to work around an issue where gMSA accounts for scheduled tasks must wait for Winlogon, W32Time and the AD controller connectivity to become available. This was done with ```-RandomDelay``` property on the task.

As far as I can see this property does nothing even though its declared in the W32 API (https://docs.microsoft.com/en-us/windows/win32/taskschd/dailytrigger-randomdelay) so was replaced with a fixed delay that simply waits 5 minutes after startup before the whole show kicks off.