Merged PR 39430: Merge order-the-right-way to master
Ensure the state file collection is ordered consistently in the code base.

It would be nice to bring in a domain type like SortedList<T> to handle this but all BCL types that deal with sorting also enforce equality and do not allow duplicates so we have to sprinkle the domain rule throughout the code base which is not ideal.

Ordering for humans (display)

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/39430/attachments/image.png)

And the system selecting a file.
Note 1159612 is the highest Id in the group for Services.Applications.FirmControl

```
Tracking 12 input files for Services.Applications.FirmControl
         State file 59dfea61-c1d7-4d9a-9362-33c537c4148a tracked files are up to date: True
         Using state file: 59dfea61-c1d7-4d9a-9362-33c537c4148a -> 1159612 -> \\aderant.com\expert-ci\prebuilts\v1\c1\cc0364c0d7a6ce509ac5b627958cc9f699fa63\1159612:Services.Applications.FirmControl
         All tracked files are up to date: C:\Source\ExpertSuite\Services.Applications.FirmControl
```