Merged PR 39434: Clean up build environment before running a build
It has been found that the model of cleaning on failure has some drawbacks. Often you want to inspect state upon a failure so cleaning up and removing assets, files and services is counter to that requirement.

This change creates a new Task on each agent that will be called before a build runs to clean up the environment. The task is used over calling as script a a particular location to provide weak coupling - if the task fails or doesn't exist no harm is done.

New task
![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/39434/attachments/image.png)

The modified Expert build definition

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/39434/attachments/image%20(2).png)