Merged PR 40424: Performance improvements for the C# analyzer.
Performance improvements for the C# analyzer.

We now no longer hammer the disk looking for csproj files to locate the GlobalSuppression file but instead query the Roslyn model for this file.
Added detection of code using SqlParameter to avoid flagging those methods as vulnerable to injection.
Upgraded dependencies to stay current and converted the analyzer to an SDK project.
#260462

Before this code was invalid.

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/40424/attachments/image.png)

Now this code is valid

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/40424/attachments/image%20(2).png)

Related work items: #260462