Merged PR 40640: Teach the build to not inspect the Git history. This speeds up local builds
Teach the build to not inspect the Git history. This speeds up local builds with 100+ commits as walking the Git history is quite slow.

If you have a local branch with 100+ commits then bm takes a while to get going as we need to find the common ancestor for build caching. Unfortunately libgit2sharp is really slow at this, so -NoBuildCache now disables looking at the Git history.