Merged PR 40845: New WPF rules. Port some tests to "VerifyCS" pattern
Introduce code analysis rules to detect

- static and member initialization
- wpf applications not using ApplicationDefinition (Types deriving from AppShell also need to use ApplicationDefinition to generate the static Main() entry point)

#261829

The "performance monitoring" initiative that caused apps to move away from ApplicationDefinition will be broken by this new requirement. I don't feel this is a great loss as no workflow exists around monitoring the data that is collected. Constantly visiting each executable is a real chore and there is not need for custom code to exist in each executable, most of the inputs can be declarative and the 'performance monitoring' feature can be integrated into the base class (AppShellApplication)

Code like this in AppShell applications is now forbidden as it has side effects.

```
  public partial class App {
        private static readonly DialogService dialogService = new DialogService();
}
```

The fix is to place initialization code in ```OnStartupAsync()```

Additionally the some tests were migrated to the recommended VerifyCS pattern https://github.com/dotnet/roslyn-sdk/tree/main/src/Microsoft.CodeAnalysis.Testing

![image.png](https://tfs.aderant.com/tfs/ADERANT/44f228f7-b636-4bd3-99ee-eb2f1570d768/_apis/git/repositories/a0f7d5b4-270c-450d-9f90-d8b8e31d4a2e/pullRequests/40845/attachments/image.png)

Related work items: #261829